Create index.html
Toast is mostly used for a show message on website users for error or success purposes. So here made simple and attractive Toast Messages Using CSS and HTML tags.

In this Toast message, we used simple icons and HTML tags for heading and for descriptions used paragraph tag. Also, we used CSS flexbox properties and for responsive we used CSS media quires.

Make it yours now by using it, downloading it, and please share it. we will design more elements for you.